Move confusing metadata init code
There were no tests covering this initialization and CI actually passes
if it is removed (see PS1), however, there are places in the code
dealing with these fields in ways that don't look safe. Instead of
removing it, this patch moves it to a more suitable place as the comment
suggested.

Bug: T282697
Change-Id: I7a6549c2d6c74a7e41901c184d0298acdeef28a2