Rename PropertyRdfBuilder to PropertySpecificComponentsRdfBuilder
In a followup I hope to create a PropertyRdfBuilder that is similar
to the ItemRdfBuilder. i.e. a composition of all the RdfBuilders
required to create the Rdf for a Property and not just the property
specific parts.

This name while rather long hopefully captures more specifically
which bits of the Rdf this builds.

Bug: T282685
Change-Id: I3f8803d764b406ed406c08713fe500d470d98709