repo: Use Item's concept URI for prefix of unit converter
We don't use anything beside "1" and items for units, the unit storage
keep them as "Q1234" which the value for units is stored as
"https://www.wikidata.org/entity/Q1234", this works in wikidata as the
item's concept uri is the same as repo's concept uri. This is not true
in commons and breaks conversion because it can't remove
"https://commons.wikimedia.org/entity/" from the URI pointing to
wikidata.org and even if it could, it would prefix it with commons which
would produce wrong RDF.

Tested in mwdebug1001 and confirms that it produces psn: values now.

Bug: T281468
Change-Id: I6d62060aaf75d7e145ab64f833b443b94397328e