Introduce client service for client DomainDb abstraction
A service is a common pattern to access functionality and to abstract
the particular construction implementation details away from the places
that just need the functionality.

This is part of the refactoring effort to have some layer of abstraction
on the database usage between Wikibase and MediaWiki.

In particular the DomainDb and associated classes and the services where
it is constructed should be the only places that know about LBFactories
and LoadBalancer classes.

Bug: T283245
Change-Id: Ic1ca0d1ac8618198eaf4b82f6d674e883db72a63