Introduce EntityStubRdfBuilder
An interface to be implemented by rdf builders of stub data
that only requires the id of the entity to be added.

This allows the implementations to choose how to get the entity
data they require rather than needed the entire entity to be passed
as is the case with EntityRdfBuilder implementers.

This is useful because typically Stub data does not require most of
an entity but just things like: labels, descriptions or entity type
so using an whole entity is excessive.

Bug: T282686
Change-Id: I3fff2e061153dc3a8464ada569e5aa37001ea0ba