Set default loadgroups for DomainDb instances
They can still be overridden by the actual calls to ConnectionManager
public methods and we can't change that. This will have to be caught in
code review.
The alternative of only exposing proxies for the ConnectionManager
methods does not seem feasible, in particular when considering that we
might also need to support the ConsistentSessionConnectionManager with
its own set of methods.

It doesn't seem possible to write tests for this behavior without
somehow making assertations about the internal properties of objects or
also implcitly testing the internal behavior of ConnectionManager.

Bug: T283245
Change-Id: Idef88bad9da626b5794ca11b0b420fa7fc319958