Add ItemStubRdfBuilder
Covered with a simple unit test that is very permissive
and simple expects the TermLookup to be used.

Also covered with an integration test inspired by
the one that currently covers the addEntityStub method
on the TermsRdfBuilder

An in memory implementation of a PrefetchingTermLookup
similar to the FakePrefetchingTermLookup is introduced
to avoid creating a very heavy mock but for use in tests
only.

Change-Id: I6efabfe4b2b0ccbab962a9091f1bd4cdbd80845b