DomainDb: don't break client repo service if repo is local
\Wikibase\DataAccess\EntitySource::getDatabaseName can be false if the
local wiki is also its own repo.
The repo is resolved to its proper name in the service, because we would
like to get rid of the `string|false` type at some point. That way, the
"DomainDb"-package can keep explicit type-hints.

getDBLoadBalancerFactory can't be mocked in a test, so we have to make
the mock inside ServiceWiringTestCase behave a bit more like the real
thing out of the box.

Change-Id: Ia382b3839960d2276ab890ff8cdec65cfd3253f3