Drop rebuilding property info table from schema updater logic
The reasoning is a bit complicated so bear with me.

This code path is only accessible if wb_property_info table doesn't
exist meaning it's either a fresh installation or it's being upgraded
from an old version of Wikibase (this table was introduced in
I2c6f873dd854 in June 2014).

 - For new installations, you don't need this rebuild as there's no
   property to have info on.
 - For old versions, we don't support such upgrade anymore. update.php
   would just refuse to update from a mediawiki that old (See T259771
   for more information).
 - For really extreme edge cases of importing the properties first and
   then installing Wikibase (which you shouldn't), there is a maintenance
   script to run.

So having this rather large code sitting here seems pointless, let's get
rid of it.

Change-Id: I169c8b25bc2e796232d681688673148dcbaeda63