Change RepoDomainDbFactory::newForEntityType to newForEntitySource
Using the entity source instead of the entity type is more future-proof
since we expect there to be multiple entity sources per entity type in
the future, so we would no longer be able to uniquely be able to map an
entity type to a database. We also already ran into a situation where
only the entity source was available, but not the entity type.

See:
https://github.com/wikimedia/Wikibase/blob/024a06a3afe18104b381a0c47f334166b0520869/repo/WikibaseRepo.ServiceWiring.php#L1759

Bug: T283342
Change-Id: I0f45d1b64e2f2cc3920591e6c92cac23f6758bab