Rewrite SerializationModifier to be more efficient
This backport combines three commits on the main branch:

* change Ifb94e358f0 (commit 786be05a0d)
* change I4cf5f76693 (commit ad4648e550)
* change If074b9cfba (commit 600331949b)

Together, they change the interaction between ResultBuilder and
SerializationModifier so that SerializationModifier is called with a
list of modifications, instead of once per modification, and only walks
the serialization array once. This is much more efficient, especially
where ResultBuilder has many modifications starting with */*/ (for
claims in subentities).

Change-Id: I5245ea3f2ec1bd11e5cdf17e7469b16559717114