Make SerializationModifier take multiple callbacks at once
This changes the interface of SerializationModifier so that it takes a
mapping from paths to callbacks, rather than only a single path and
callback, and can therefore apply many modifications at once. The
implementation isn’t touched yet: for now, the new interface is
implemented as a simple loop over the paths and callbacks, calling the
existing implementation each time. This will be changed in a separate
commit.

Change-Id: Ifb94e358f089ba90102f071c02ca34e3274b8944