Inject RepoDomainDb into SiteLinkTable
The code to create the RepoDomainDb in DirectSqlStore is fairly ugly,
but as far as I can tell we can’t use RepoDomainDbFactory because we
don’t have an entity source there. (Hopefully T283334 will improve on
this.) Similarly, in rebuildItemsPerSite.php we use newRepoDb(), which
is discouraged, but we don’t have an entity source there either.

Bug: T284651
Change-Id: I3d6e69bb92fa3c1607e3eb70625346949b4fbd73