Reimplement SerializationModifier
Instead of iterating over the paths and callbacks and recursing through
the array each time – which is especially inefficient if the paths
contain segments like */* – first prepare the callbacks into a more
convenient format, then iterate the array *once* and process all the
callbacks together.

I locally tested the performance impact of this on two levels:

- A PHPUnit test that downloads the Q42 entity data (in PHP format),
  then calls getEntitySerializationWithMetaData() on it 1000 times.
  Before this change, it took roughly 32-33 seconds; with this change,
  only around 6 seconds. (Specifically: on master, 32056ms, 31802ms,
  31983ms; on change Ifb94e358f0, which changes the calling convention
  but still uses the old implementation, 33182ms, 33047ms, 35243ms; with
  this commit, 6128ms, 5844ms, 6184ms.)

- A curl command that makes 1000 wbgetentities API requests (XML format)
  for the largest entity I have locally (according to Special:LongPages)
  and throws all the results away. On master, the 1000 requests took
  2m5.7s; with this change, 1m55.5s.

As a bonus, this implementation is also quite a bit shorter than the
previous one.

Change-Id: If074b9cfbab278221439535467f40f501e13ff35