Make SerializationModifier work in any order
Previously, you had to take care to pass callbacks into
SerializationModifier in the right order, with the least specific
callbacks at the end, so that the more specific callbacks didn’t iterate
over keys added by the less specific callbacks. Avoid this by sorting by
the number of slashes in the key (more slashes first).

In theory, this removes some flexibility from callers, but in practice
nobody seems to have needed that flexibility and this should make it
harder to accidentally misuse the class.

Change-Id: I4cf5f76693935a12a8b129f0ef20dd59c584e5e1