Add documentation for API param values
The descriptions for the aspects are the same for both endpoints
wbentityusage and wblistentityusage and so rather than duplicating the
copy and making the translators do double the work, we're just reusing
the copy from wbentityusage for wblistentityusage as well.

The description for aspect O differs from the description for that
aspect in other places because the documentation in those other places
is wrong/outdated. It will be updated in other commits, etc.

Bug: T283040
Change-Id: I9b6ebb7c82ae7b847ffcda0186933cb68e29eaf1
(cherry picked from commit 645f597e90a8e77d17e6f3341833cb2f7671ea0c)