Slightly refactor complex loop to appease phan
In the following commit type hints will be added that caused phan to
complain about `$prRow` being null in the `formatPageData()` method call.

So this commit refactors that loop slightly to work around it.
`$currentPageId` was actually the pageid of the `$prRow` variable,
which means "previous row" (I think).

It is separate from the following commit to make both easier to review.

Also, this fixes a bug in that loop that would cause the last item in
the list to have wrong and inconsistent pageid and label, if the
response had to be paginated and that pagination would happen exactly
between two pages.

Change-Id: I54799eb5f84d1aa2da00982042ba01ff77b0d395