bridge: retry initialization once
If we get a badtoken error from CentralAuth, assume that there was a
race condition and retry the initialization; if that succeeds, track the
error as a resolved one, rather than a regular error.

Tested with the following snippet in LocalSettings.php:

    $wgHooks['ApiCheckCanExecute'][] = function ( $module, $user, &$message ) {
        if ( $module->getModuleName() !== 'wbgetentities' ) {
            return true;
        }

        $good = file_exists( '/tmp/ApiCheckCanExecute' );
        touch( '/tmp/ApiCheckCanExecute' );
        if ( !$good ) {
            $message = [ 'badtoken', 'apierror-centralauth-badtoken' ];
            return false;
        }
        return true;
    };

This fails the first wbgetentities request but allows subsequent ones.
Delete /tmp/ApiCheckCanExecute to reset.

Bug: T285064
Change-Id: I4fbf76c658dbb3360553e94797bfe00d3902fd4d