FP: Make TypeDispatchingExistenceChecker Type and Source Dispatching
In order to allow both local and federated entities of the same type to
exist on a Wikibase, we need dispatching services that delegate to
either a DB-backed or an API-backed service depending on the type of
entity source.

We previously swapped out the DB-backed services for
API-backed ones for Properties when Federated Properties was enabled,
but this is no longer sufficient.

This patch turns TypeDispatchingExistenceChecker into the first service
that dispatches by entity source and entity type.

Bug: T285303
Change-Id: Iafba07ea7fcf785b96fb37b38f741efbacc8913d