Update MockRepoClientCentralIdLookup to UserIdentity (part 1)
Given that CentralIdLookup is only extended in CentralAuth, maintained
by the platform team, and in a Wikibase test, we've made a decision to
change its signature from User to UserIdentity without deprecation.

This is part 1 of the update, removing the old parameter type and adding
return types. Part 2 will add back the new parameter type.

Needed-By: I67bf4f2442269b8aa1002868ede9a71a5ca43a84
Change-Id: I4a08af2227d0dfd833e16469df3f76e397dce669