bridge: use errorformat=raw for all API requests
Previously, we used it for some API requests, but others used the
default error format. Let’s always use the “raw” format so that we can
always consistently inspect the errors. (We never show the API error
messages to the user, so we don’t need a format where the text is
rendered as a proper message.)

Also, make the order of API parameters more consistent: errorformat and
formatversion towards the end, and same order in src/ and tests/.

Bug: T285064
Change-Id: Ic11a10708b551c01524e9d5040a2e3d55de278b5