FP: Introduce EntitySourceLookup
This class provides a way to figure out which
source a given EntityId has come from.

To do this it looks in the EntitySourceDefinitions
for suitable EntitySources. i.e. ones that match
the conceptBaseUri and are an API source or
a Database EntitySource that supports the type of
Entity the EntityId is.

In order to differentiate between EntitySources from
DBs or APIs a 'type' is added to EntitySource.
Note: this patch doesn't allow creating non-db EntitySources
This is introduced in a follow-up

Co-authored-by: Jakob Warkotsch <jakob.warkotsch@wikimedia.de>
Co-authored-by: tzhelyazkova <tonina.zhelyazkova@wikimedia.de>
Bug: T285455
Change-Id: Ie4418cf50ccb74cefa64640d3d86ed950a22fb1c