Make TemplateModuleTest work with strict_types
PHP doesn’t support clearing the stat cache for only one file.
clerastatcache() does take a $filename argument, but only after a
$clear_realpath_cache flag (type bool – we were passing the string $file
to that parameter, causing a type error), and it only affects the
realpath cache, not the rest of the stat cache. Let’s just clear the
whole stat cache.

Change-Id: Ifddca6a99dfd0bf7a26132ac0a17dee1d3b15fae