Make RepoHooksTest work with strict_types
EntityHandler::PARSER_VERSION is an integer, so passing it into
preg_quote() requires an implicit cast that’s illegal when strict_types
is enabled. While I admire the dedication to using preg_quote() when
building regexes, in this case I think it’s excessive, since the quoted
value can’t actually contain any metacharacters (even after being cast
to a string), and the regex is also only used in a test. Removing the
cast seems to me like the easiest solution.

Change-Id: Iae5339f4ab845872722e7a44b4a71c2b8150ee32