Make TermTestHelper work with strict_types
$length / strlen( $text ) can be a float (e.g. with the default
multilang limit of 250 and the default $text of "Test" – 250 is not
divisible by 4), and in that case passing it into str_repeat() requires
an implicit cast to int, which is not allowed with strict_types enabled.
Use ceil() to ensure an integer result with an appropriate value.

Also, document the function – previously, it was (in my opinion) unclear
if the string was “overly long” subjectively, but had the exact length
specified in the parameters, or if it was “overly long” compared to that
length (i.e. was guaranteed to be longer).

Change-Id: I610a341ce06b4687fad0df6e9b91b14c5f90744a