Prefetch Item StubEntityData
Introduce new interface for stubRdfBuilders with a method for
prefetching data that the builder may need.

Call this method for all entities that need resolving.  Also call it
(potentially a second time) for addEntityStub, since this public method
may possibly be called from somewhere which didn't also mark these
entities for prefetching.

Bug: T281272
Change-Id: I66435aa931a1ce4b2b351d2b8e00e964572a13c7