EntityDataRequestHandler: Remove ob_clean() call
ob_clean() was added in 2013, in Ie9f2f8944a. It doesn't seem to be
necessary here and it breaks the Apache set up used in Quibble as well
as in MediaWiki-Docker, and is reproducible with fresh-node as well.

Bug: T282154
Bug: T276428
Change-Id: Ia1126bbbabada07f7651dd59361aff667fe1d5fd
(cherry picked from commit f78be2bc9fd24a68aa007acfd9260b21fcdf359f)