Remove WikibaseRepo foreignRepositories setting
With this change, 'entitySources' is the one and only source of truth
for the configuration of, well, entity sources. Previous legacy
settings, such as 'foreignRepositories', 'entityNamepaces', and others,
are no longer supported.

Previously, there was some quite special support for 'entityNamepaces'
in particular. WikibaseSettings would run two hooks to populate it,
WikibaseRepoEntityNamespaces and WikibaseClientEntityNamespaces, where
other extensions could register a namespace for their entity types.
(WikibaseSettings called the hook with the original 'entityNamepaces'
value, so if the extensions didn’t check for any existing configuration
in the array, then that would simply be lost, as far as I can tell.
Neither WikibaseLexeme nor WikibaseMediaInfo paid heed to any existing
configuration in the array.) Then, it set the 'entityNamepaces' setting
to the modified array; it also looked through the 'repositories' setting
and added any 'entityNamepaces' not mentioned in any existing repository
to the local repository (since WikibaseRepo doesn’t use 'repositories' –
it has 'foreignRepositories' instead – that last part only affected
WikibaseClient). Finally, WikibaseRepo would parse entity sources out of
that 'entityNamepaces' setting and others, but only if 'entitySources'
wasn’t configured (if it was, the hook results weren’t used).

This commit removes a lot of that special handling. We still run the
WikibaseRepoEntityNamespaces hook, because we need a default entity
sources configuration that includes extension entity types for various
tests, but we don’t do it in WikibaseSettings anymore – instead, the
hook now runs in the example 'entitySources' callback, and is therefore
only intended for default / suggested entity namespaces. If a user
configures their own 'entitySources' (which is now required if the
example settings aren’t used – the default just throws an exception),
the hook isn’t even called, and entities can be stored in whichever
namespaces the user prefers. (The WikibaseClientEntityNamespaces hook,
on the other hand, is dropped – a non-local client wiki must configure
custom 'entitySources' anyway.)

The FederatedPropertiesEntitySourceDefinitionsConfigParser is also used
differently. Previously, it was used only when 'entitySources' were not
configured, with the effective entity sources parsed from the legacy
settings. I assume the expectation was that this would have mostly been
the default settings – as far as I can tell, the federated properties
parser discards any non-local entity sources (which may have been
configured via 'foreignRepositories'). Now, we call it when there is
exactly one entity source. It doesn’t matter whether that one entity
source comes from the default config or was manually configured – in
either case, the federated properties parser can split properties out of
it and return two entity sources, the original one and a new one for
federated properties. On the other hand, if the user configured more
than one entity source, then they’re responsible for setting up
federated properties in one of those sources.

Bug: T285472
Change-Id: Ib1a66f95074de7de7c4c16c78de236d4b00399b4
Co-Authored-By: Lucas Werkmeister <lucas.werkmeister@wikimedia.de>
Depends-On: I82582576193d13d983717347208e131b7d342ff1