docs: Treat table names as code
This makes both the rendered and the unrendered markdown much easier to
read.

Script paths should also be printed as monospaced code, but somehow
Doxygen has some magic to link them to their respective file
documentation directly, which would seem to preferable. This doesn't
happen for the core script `update.php`, so that is wrapped in
backticks.

Unfortunately, [allowArbitraryDataAccess] = true can't be changed to
monospace as it should be, because [allowArbitraryDataAccess] is a link.

Change-Id: Ie010fbad058870206d6ccc8dafb291c73a3975f7
(cherry picked from commit 788702afebfa8d3e326ff4a6005f0c5138722506)