Don’t count self in NoBadDependencyUsageTest
Previously, NoBadDependencyUsageTest’s counts of allowed references
included the strings in its own file: notably, the allowed count

    'Wikibase\\\\Repo\\\\' => 1,

was really just a match for the previous line

    'Wikibase\\Repo\\' => 1,

and nowhere else in Wikibase has a reference to that string with double
backslashes. Also, the strings 'WikibaseRepo::' and 'WikibaseClient::'
were split up into two parts so that they wouldn’t match themselves.

This is all fairly confusing, and I think it’s better to just skip
NoBadDependencyUsageTest.php when counting usages. This unfortunately
requires the use of getRealPath() (resolving symlinks etc.), but the
slowdown seems to be tolerable.

Change-Id: I17a25250250d817e6724bff1f5e0e027d19309eb