Rewrite NoBadDependencyUsageTest as NoBadUsageTest
This makes the test much more fine-grained: instead of just having a
single number of occurrences for each bad pattern, across all of lib/,
support specifying the number of allowed occurrences for each directory
or path. Also, extract the whole implementation into a base class that
can also be used in repo and client later.

Bug: T286209
Change-Id: I01c9526240ffa0ac1a64b8b527dba707110a8834