Remove WikibaseClient repositories setting
This removes the default repositories setting, along with support for
parsing this and other “legacy” settings into entity sources
(implemented in EntitySourceDefinitionsLegacyClientSettingsParser).
Instead, there is now a default entitySources setting, and clients of
non-local repos must configure custom entitySources.

The default entitySources, like the default repositories before them,
copy the current repo configuration if the current wiki is a repo (and
throw an exception otherwise). However, instead of directly copying the
repo entitySources *setting* (in the old code: the entityNamespaces
setting), we build an entitySources array from the repo’s effective
EntitySourceDefinitions *service*, which may have been loaded from
legacy settings. In theory, this should also give us the federated
properties entity source if federated properties are enabled in the repo
configuration of the wiki; however, the WikibaseClient service wiring
for EntitySourceAndTypeDefinitions can’t parse the “api” entity source
type yet (only the corresponding repo wiring supports it), so for now we
filter out non-“db” entity sources.

This is a prerequisite for removing the entityNamespaces setting from
the repo (we can’t remove it there while client still tries to copy it
for its own config).

The default itemAndPropertySourceName is also changed to read the source
name from the repo local entity source (if the wiki is also a repo),
since the copied entitySources may not include a 'local' source.
(Without this, EntityHandlerTestCase fails, because it configures a
local 'test' source.)

ClientDefaultsTest is reenabled. It’s been skipped for two and a half
years – if it’s still flaky after all this time, and the major changes
of this commit, then we should look into that, and not just hide the
problem, in my opinion.

Bug: T285471
Change-Id: I4243d2a00d4ec9ebf5db5d70d8589f219649a92c