Remove default repositories for non-repo clients
I don’t know if this ever worked very well (assuming that the hard-coded
Wikidata defaults were what you wanted in the first place), but I
believe this must have been broken for a while: the default repoDatabase
for clients without a local repo was `null`, and the EntitySource
constructor will throw an exception in that case. (And I don’t see any
code in EntitySourceDefinitionsLegacyClientSettingsParser to turn that
`null` into an allowed value).

Bug: T285471
Change-Id: I1093313eebd6e5e03e81488a4e405c008e194ebc