Add tags parameter to wbmergeitems API
This adds a tags parameter to MergeItems and then adapts all the various
levels between it and WikiPageEntityStore::saveEntityContent() to pass
through the list of tags. Validation (is the user allowed to apply these
tags?) happens when MediaWiki parses the parameter (TagsDef::validate),
we don’t need to deal with it directly. (I didn’t add a test for this,
because MergeItemsTest is currently not implemented as an ApiTestCase,
which means it doesn’t do parameter validation.)

The signature of EntityRedirectCreationInteractor::createRedirect()
requires an incompatible change (add $tags between $bot and $context),
and that method is called in WikibaseLexeme, so temporarily support both
the old and the new signature.

Bug: T286778
Change-Id: Ib2b03a50558230f0e7ad17cb79a9d1445926c9d8
Depends-On: I027edaf1aa4519aa5fbd89401b7330fdfc3fd09f