Tag Termbox edits
This introduces a new server-side configuration setting for tags that
should be added to edits made with Termbox v2. To send the configuration
to the client-side code without introducing a new ResourceLoader module
(mw.config.values.*), turn the termbox module into a package files
module [1] and add a synthetic config.json file to it with the tags
(and, possibly, more config in the future); termbox has been taught to
require( './config.json' ), which means that the synthetic file needs to
be in dist/ so that it’s next to the built JS code.

[1]: https://www.mediawiki.org/wiki/ResourceLoader/Package_files

Bug: T286775
Change-Id: I097c5c573413792e0dad32760d34632d5475a08e