Remove pointless @see tags in RepoApi
Some of the methods tagged @see post didn’t even use post(); and in
general I think these tags are useless and make it more annoying to find
actual uses of post(). The usage of post() by these other methods is
purely internal – the documentation of post() is not helpful to callers
of the other methods (there is not, say, a parameter that’s passed
through from one method to the other, and which is only documented in
post(), which would justify the @see).

Change-Id: I0239e202ebcf924b43dc13277c79627fea6ff0ae