Handle unknown flavor errors better
Introduce a dedicated exception type that can be thrown in the two
places that currently translate a flavor string into flavor flags, and
catch that exception in dumpRdf.php and in the EntityDataRequestHandler.
In dumpRdf.php, we can just show the untranslated message as generated
by the exception class; for Special:EntityData, introduce a new
interface message to be used.

Bug: T286275
Change-Id: If59fad8b2b93a7f3a8f1a2830554dcbe44c9ca7d