Add tags to edits made via linkitem
The tags configuration is added to the config.json file of the init
module, which also contains the site information. Strictly speaking,
neither piece of information is needed until the link has been clicked,
so it could also live in the non-init module, but I don’t want to deal
with the extra difficulties of that at the moment.

Because the contents of that file are cached, the client-side code has
to deal with the possibility that the tags member isn’t present (i.e.
use linkItemConfig.tags || [] instead of just linkItemConfig.tags); and
since we already have to do that, I figured we might as well only add
the tags to the value in PHP if they’re configured, and otherwise leave
them out completely, and save a handful of bytes.

Bug: T286771
Change-Id: Ic7ef0830e4f5a6edbdc817ec35fe41c5341a2229