Introduce SnakNormalizer
This takes the normalizer-factory-callback from the data type
definitions, determines the right normalizers to use according to the
property data type and data value type, and normalizes the value of the
snak (assuming it’s a value snak, of course).

If the property data type can’t be determined, a message is logged, but
only at info level: warning or higher wouldn’t be appropriate, since
users can trigger this message by sending snaks with nonexistent
property IDs to the API.

In the tests, we use normalizers that compose nicely in a commutative
manner: we don’t care whether the value was first uppercased and then
doubled or the other way around.

Bug: T251480
Change-Id: I6ed153dc4d402ee2afcf928449cbac7137abd454