Remove wfWikiID() fallback from PidLock
The function has been deprecated for a while. It was already replaced
with WikiMap::getCurrentWikiId() when pruneChanges.php calls the
constructor, in change Ib5da5fe23a (commit dc257a39b1), but PidLock
still had a fallback to wfWikiID(), which is never actually used (the
constructor is only called once, in pruneChanges.php). Let’s just remove
that and require all callers to specify the wiki ID.

Change-Id: I15f05c4003917411c753d1ad77d4e11d4f4653ef