Use wfArrayPlus2d() to merge definitions arrays
array_merge_recursive() merges more levels than we want to, and will do
the wrong thing if the same key is present in both arrays. It’s safer to
use wfArrayPlus2d() instead. Since wfArrayPlus2d() will, in case of key
conflict, keep the value in the earlier array, we swap the arguments so
that the more specific array (repo/client) comes before the general one
(lib).

Bug: T287139
Change-Id: I68a12d4c2e3e2b0905c6fb76d29f2908e15b30cc