Return normalized snaks from SetClaim, SetReference
All API modules editing snaks should return the normalized version of
the data they’re adding. Most do this already, due to how they get the
modified data; only wbsetclaim and wbsetreference were different. In
SetClaim, it’s fairly simple to get the statement with the same GUID out
of the patched entity; in SetReference, we unfortunately don’t have a
stable identifier for the reference being set (only its hash, which
changes under normalization), so we work around this by cloning the old
reference list and checking which of the post-patch references isn’t in
it, assuming that this is the one reference that was changed and should
be returned. This seems better to me than adding some getReference()
method to ChangeOpReference. (Cloning a ReferenceList is allowed, the
Statement::__clone() implementation does it too.)

Tests will be added in a separate change (I592a76ba4e), since we should
really test not only the two API modules that are fixed here, but also
the ones where it happens to work already.

Bug: T289501
Change-Id: Ib3eea824dc9cf9aa623b60d450a6251e93b3c51b