Test that APIs return normalized data values
None of the currently implemented normalization is suitable for these
tests: commonsMedia normalization is skipped in tests (and we don’t want
to make the tests contact production Commons either), and Unicode
normalization is applied to all API responses anyways, whether we do it
at the individual API module level or not. The simplest solution, I
think, is to create a new, test-specific datatype with custom
normalization and use that. Since it’s needed by all the tests, we can
put it in WikibaseApiTestCase; the normalization chosen is “convert
string to uppercase”, which is nice and simple.

Bug: T289501
Change-Id: I592a76ba4e1586071514fa8fac877d6a1261224b