selenium: Update README.md files
This commit continues work started in 4283908.

Simplify and update README files in repo/tests/selenium and
client/data-bridge/tests/selenium using the https://phabricator.wikimedia.org/P8206 template. The template
is created to provide minimal information on how to run tests that is
still useful. The readme should be short and provide links to more
information.

Bug: T282237
Bug: T289853
Change-Id: I4603c386a1637fbad8b74a975f08e0fa44a2772e