FP: use dynamic IDs in tests to avoid collisions
This test fails with the following error when run in a suite with other
tests that create Items and Properties because it explicitly defines
IDs before creation:
```
Wikibase\Lib\Store\StorageException: Could not create a new page.
It already exists.
```

We can let the IDs get set when the Entity is saved as it would happen
in a production scenario.

Change-Id: I4e5a31a78b5ce920d0f3227da4b6ccef01556b7d