tests: Remove use of AtEase
Something somewhere is leaving error_reporting in a dirty state
causing AbuseFilter's ConsequencesExecutorTest case to fail for
the core change Ic9fee6cdd88001025.

Wikibase is one of the CI dependencies of AbuseFilter.

Per T253461, we're meant to eventually remove these anyway, so might
as well get it over with avoid its state problems, which are especially
fragile during tests where exceptions are not uncommon, thus not
reaching the restore call.

Change-Id: I2a665f09a357f2f2cc258d8c4011d49a7ab9c13b