Fix secondary CI emails
If the HTML body is specified in a separate file, then it won’t get any
variables substituted, so let’s specify it directly in the YAML file.
Also, the syntax to substitute most of the variables was wrong.
Apparently, the GITHUB_* environment variables aren’t available in input
parameters (“with:”) [1]; we can replace most of them with values from
the github context [2] instead, but the GITHUB_SERVER_URL is
inexplicably missing there, so let’s just hard-code that.

[1]: https://github.community/t/how-to-pass-environment-variable-to-an-input/16277
[2]: https://docs.github.com/en/actions/reference/context-and-expression-syntax-for-github-actions#github-context

Bug: T291250
Change-Id: I2a777a0aabad516cdc73fd39aedb7617e3cceb87