Prevent creation of true duplicates
We use memcached (or any other BagOStuff implementation) as a simple
lock via the “add” command, “locking” each sitelink before saving an
item, and “unlocking” it again as soon as it’s been written and
committed to the proper site link store. (A previous change, change
I3f458bd343, ensures that SqlSiteLinkConflictLookup can read sitelinks
from the SQL store as soon as they’ve been committed, even if that
commit was after the beginning of our transaction.)

This is implemented as a new SiteLinkConflictLookup implementation,
which also writes to the cache if DB_PRIMARY was specified. The
memcached implementation may not be able to return the conflicting item
(if the add fails, but the subsequent get returns null), but the
contract of SiteLinkConflictLookup and UniquenessViolation was already
adjusted for this in change I7c75fe6895.

Bug: T291377
Change-Id: I3f458bd3434341c6bdb90e6d2f4999c85c46dc14