Entity URLs: force english namespace names for interwiki links.
Interwiki Title objects should always use NS_MAIN, and have any
namespaces be part of the title text. The reason is that we don't know
the target's content language, so we don't know the local name
of the "Special" namespace. The only names that are certain to work
across wikis are the English ones.

Change-Id: I54bf684be633a12210bbf115474a03de3b92cbfa
Needed-By: I3d5b834a71c0745596a7f4dc914fb7a61e026060