Allow unknown item ID in SiteLinkConflictLookup
An upcoming SiteLinkConflictLookup implementation (see change
I3f458bd343) won’t always be able to return the conflicting item ID, so
let’s make it optional in the SiteLinkConflictLookup conflict array and
in UniquenessViolation::getConflictingEntity().

As far as I can tell, all callers are compatible with this – the main
one is ChangeOpsMerge::removeConflictsWithEntity(), which passes the
conflicting ID into EntityId::equals(), where null is allowed.

Bug: T291377
Change-Id: I7c75fe6895554f146eaf628fb400886867da5663