Remove useless code comments and unreachable statements in examples
These example configs some are already in the documentations and some
are very specific to mediawiki itself (and possibly outdated) which the
maintainer should look at mediawiki documentation and not us.

Our official stance these days is that we offer releases and docker images
which handles this out of the box making this bits even more useless.

Bug: T291617
Change-Id: Ia8f00f33c78aa7f73c41a4c1b3544d77b0351e1d