Remove some duplicate property label validation
Given the new LabelUniquenessValidator, we can now remove some of the
validation that was already in place. In FingerprintUniquenessValidator,
all the property validation can be removed. In SpecialNewProperty, it’s
technically also redundant, but we keep it so that the error can be
detected before we’ve assigned (consumed/wasted) a property ID.

Bug: T289473
Change-Id: I9f18b9daaf62260353045d0b27ddddcc06d0d697