Only add labels to entity links in comments and special pages
Previously, HtmlPageLinkRendererEndHookHandler tried to determine
whether a link appeared in a comment based on the current action
(history or view+diff), but this was fragile and meant that labels would
be added even to links in the page content on diff views, which we don’t
want. Instead, we can use the newly introduced “comment” flag on
LinkRenderer to get this information reliably.

Bug: T292203
Change-Id: I4b55cf8f54fb13c4ef1c50cb26ac133170a6a129
Depends-On: I9164f760e8b46e05218fb36f430bc36ef1fdf30f